Use int as data type for the array. No other quanity and Phase uses nearest neighbor interpolation anyways.

+++ interpBilinear.h
@@ -30,7 +30,7 @@ double backwardInterpolateDensityBilinear(double T, double z, int nT, int nRho, 
 	double* TAxis, double** uArray);
 double interpolateValueBilinear(double rho, double T, int nT, int nRho, double* rhoAxis,
 	double* TAxis, double** zArray);
-double interpolateValueNearest(double rho, double T, int nT, int nRho, double* rhoAxis,
-        double* TAxis, double** zArray);
+int interpolateValueNearest(double rho, double T, int nT, int nRho, double* rhoAxis,
+        double* TAxis, int** zArray);
 int findIndex(double x, double* xAxis, int nX);
 #endif
